Clean up local datasource
