+++ impl/src/main/scala/quasar/impl/local/LocalParsedDatasource.scala
@@ -18,15 +18,21 @@ package quasar.impl.local
 
 import slamdata.Predef._
 
-import quasar.connector._
+import quasar.connector.{CompressionScheme, QueryResult, MonadResourceErr, DataFormat}, DataFormat.JsonVariant
 import quasar.connector.datasource.LightweightDatasourceModule
-import quasar.impl.parsing.ResultParser
 
 import java.nio.file.{Path => JPath}
+import scala.collection.mutable.ArrayBuffer
 
-import cats.effect.{Blocker, ContextShift, Effect, Timer}
-import fs2.io
+import cats.effect.{Blocker, ContextShift, Effect, Sync, Timer}
+
+import fs2.{compress, io, Chunk, Pipe}
+
 import qdata.{QDataDecode, QDataEncode}
+import qdata.tectonic.QDataPlate
+
+import tectonic.{json, csv}
+import tectonic.fs2.StreamParser
 
 object LocalParsedDatasource {
 
@@ -44,9 +50,51 @@ object LocalParsedDatasource {
     EvaluableLocalDatasource[F](LocalParsedType, root) { iRead =>
       val rawBytes =
         io.file.readAll[F](iRead.path, blocker, readChunkSizeBytes)
-      val parsedValues = rawBytes.through(ResultParser.typed(format))
+
+      val parsedValues = rawBytes.through(encode[F, A](format))
 
       QueryResult.parsed[F, A](QDataDecode[A], parsedValues, iRead.stages)
     }
+  }
+
+  ////
+
+  private val DefaultDecompressionBufferSize: Int = 32768
+
+  @SuppressWarnings(Array("org.wartremover.warts.Recursion"))
+  private def encode[F[_]: Sync, A: QDataEncode](format: DataFormat): Pipe[F, Byte, A] = {
+    format match {
+      case DataFormat.Json(vnt, isPrecise) =>
+        val mode: json.Parser.Mode = vnt match {
+          case JsonVariant.ArrayWrapped => json.Parser.UnwrapArray
+          case JsonVariant.LineDelimited => json.Parser.ValueStream
+        }
+
+        StreamParser(json.Parser(QDataPlate[F, A, ArrayBuffer[A]](isPrecise), mode))(
+          Chunk.buffer,
+          bufs => Chunk.buffer(concatArrayBufs[A](bufs)))
+
+      case sv: DataFormat.SeparatedValues =>
+        val config = csv.Parser.Config(
+          header = sv.header,
+          row1 = sv.row1.toByte,
+          row2 = sv.row2.toByte,
+          record = sv.record.toByte,
+          openQuote = sv.openQuote.toByte,
+          closeQuote = sv.closeQuote.toByte,
+          escape = sv.escape.toByte)
+
+        StreamParser(csv.Parser(QDataPlate[F, A, ArrayBuffer[A]](false), config))(
+          Chunk.buffer,
+          bufs => Chunk.buffer(concatArrayBufs[A](bufs)))
+
+      case DataFormat.Compressed(CompressionScheme.Gzip, pt) =>
+        compress.gunzip[F](DefaultDecompressionBufferSize) andThen encode[F, A](pt)
+    }
+  }
+
+  private def concatArrayBufs[A](bufs: List[ArrayBuffer[A]]): ArrayBuffer[A] = {
+    val totalSize = bufs.foldLeft(0)(_ + _.length)
+    bufs.foldLeft(new ArrayBuffer[A](totalSize))(_ ++= _)
   }
 }
